Improve callbacks by passing the view as a parameter
This way one can directly call the `fadeMeOut` method on the actual view instead of having to call `[TSMessage dismissActiveNotification]` (which would be necessary if you want to dismiss the view programmatically in the callback, because you have disabled manual dismiss via `canBeDismissedByUser:NO`).

I find it better to deal with the actual instance of the view rather than relying on it being the currently active notification.

Right now it is only possible for it being the currently active notification, but these changes will make sense in conjunction with the next changes I'd like to propose in the next commits. ;)