Remove iOS 6 compatibility and clean up
This is so we can go forward and remove some of the incomplete edge case handling when calculating height and offset of the message view in `fadeInNotification:`, which depends on the statusbar and navigationbar being present or not.

The demo project now works for all cases and shows the buttons to toggle statusbar and navigationbar.