Moved sound handling to TSMessage+Sounds category
Created TSMessage+Sounds category to extend TSMessages.
Category includes `TSMessageSoundPlayer` object that responds to
NSNotification.
Category header is imported inside `__has_include` macro in
TSMessageView+Private.
Added `TSMessages` as a subspec in `.podspec` and set to
`default_subspec` so its always included.
Added `Sounds` subspec in `.podspec` that includes the relevant files
and frameworks, with a dependency on the `TSMessages` subspec.